requirements.txt
# Core dependencies
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional dependencies for full functionality
# slack-sdk>=3.21.0  # For Slack integration
# redis>=4.6.0       # For caching and async processing
# psycopg2>=2.9.0    # For PostgreSQL database
# sqlalchemy>=2.0.0  # For database ORM

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# flake8>=6.0.0
# black>=23.0.0
# mypy>=1.5.0

